- kernel-install: update of parameters for jasperreport
git-svn-id: https://efaps.googlecode.com/svn/eFaps-Kernel-Install/trunk@8130 fee104cc-1dfa-8c0f-632d-d3b7e6b59fb0

+++ src/main/efaps/ESJP/org/efaps/esjp/common/jasperreport/StandartReport_Base.java
@@ -249,8 +249,8 @@ public abstract class StandartReport_Base
             final JRTextExporter exporter = new JRTextExporter();
             exporter.setParameter(JRExporterParameter.JASPER_PRINT, _jasperPrint);
             exporter.setParameter(JRExporterParameter.OUTPUT_STREAM, os);
-            exporter.setParameter(JRTextExporterParameter.CHARACTER_HEIGHT , new Integer(10));
-            exporter.setParameter(JRTextExporterParameter.CHARACTER_WIDTH , new Integer(6));
+            exporter.setParameter(JRTextExporterParameter.CHARACTER_HEIGHT , new Float(10));
+            exporter.setParameter(JRTextExporterParameter.CHARACTER_WIDTH , new Float(6));
             exporter.exportReport();
             os.close();
         }
